+++ CHANGELOG.md
@@ -12,8 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - React Router Support and basic routing
 - Home component (./src/components/pages/Home)
 - Added pages directory (./src/components/pages/Home) to aliases
+- Added font files (./src/globalAssets/fonts/)
   
 ### Changed
+- Added font support to Global Styles
 
 ### Packages Installed:
 - React Router v6.9.0